# OdhTimeseries/docs/swagger.json
{
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "A Go REST API server for managing timeseries data using Gin framework and PostgreSQL with a simplified, performance-optimized schema.",
        "title": "Timeseries API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.example.com/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/types": {
            "get": {
                "description": "Get a paginated list of all types, optionally including sensors that have timeseries for each type",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "types"
                ],
                "summary": "List all types",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Offset for pagination",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 50,
                        "description": "Limit for pagination",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Include sensors with timeseries for each type",
                        "name": "include_sensors",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of types with pagination info",
                        "schema": {
                            "$ref": "#/definitions/models.ListTypesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/types/{name}": {
            "get": {
                "description": "Get a specific type by name along with all sensors that have timeseries for this type",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "types"
                ],
                "summary": "Get a type by name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Type name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Type with sensors and timeseries info",
                        "schema": {
                            "$ref": "#/definitions/models.TypeWithSensors"
                        }
                    },
                    "404": {
                        "description": "Type not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.DataType": {
            "type": "string",
            "enum": [
                "numeric",
                "string",
                "json",
                "geoposition",
                "geoshape",
                "boolean"
            ],
            "x-enum-varnames": [
                "DataTypeNumeric",
                "DataTypeString",
                "DataTypeJSON",
                "DataTypeGeoposition",
                "DataTypeGeoshape",
                "DataTypeBoolean"
            ]
        },
...
        "models.ListTypesResponse": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "types": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TypeWithSensors"
                    }
                }
            }
        },
...
        "models.SensorTimeseriesInfo": {
            "type": "object",
            "properties": {
                "sensor_name": {
                    "type": "string"
                },
                "timeseries_id": {
                    "type": "string"
                }
            }
        },
...
        "models.Type": {
            "type": "object",
            "properties": {
                "data_type": {
                    "$ref": "#/definitions/models.DataType"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "metadata": {
                    "type": "string",
                    "example": "{\"category\":\"environmental\"}"
                },
                "name": {
                    "type": "string"
                },
                "unit": {
                    "type": "string"
                }
            }
        },
...
        "models.TypeWithSensors": {
            "type": "object",
            "properties": {
                "sensors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SensorTimeseriesInfo"
                    }
                },
                "type": {
                    "$ref": "#/definitions/models.Type"
                }
            }
        }
    }
}